Use spec attestation time verification in gRPC requests (#6429)
* Re-use attestation time verification

* lint

* fix imports
