+++ beacon-chain/sync/utils_test.go
@@ -3,13 +3,9 @@ package sync
 import (
 	"math/rand"
 	"testing"
-	"time"
 
 	ethpb "github.com/prysmaticlabs/ethereumapis/eth/v1alpha1"
-	"github.com/prysmaticlabs/prysm/beacon-chain/core/helpers"
 	"github.com/prysmaticlabs/prysm/shared/bytesutil"
-	"github.com/prysmaticlabs/prysm/shared/params"
-	"github.com/prysmaticlabs/prysm/shared/roughtime"
 )
 
 func TestSortedObj_SortBlocksRoots(t *testing.T) {
@@ -78,35 +74,5 @@ func TestSortedObj_NoDuplicates(t *testing.T) {
 			t.Errorf("Duplicated root exists %#x with block %v", newRoots[i], b)
 		}
 		rootMap[newRoots[i]] = true
-	}
-}
-
-func TestValidateAggregatedTime_ValidatesCorrectly(t *testing.T) {
-	const genesisOffset = 1200
-	genTime := roughtime.Now().Add(-(genesisOffset * time.Second))
-	currSlot := helpers.SlotsSince(genTime)
-	invalidAttSlot := currSlot - params.BeaconNetworkConfig().AttestationPropagationSlotRange - 1
-	err := validateAggregateAttTime(invalidAttSlot, genTime)
-	if err == nil {
-		t.Error("Expected attestation time to be invalid, but it was marked as valid")
-	}
-	timePerSlot := time.Duration(params.BeaconConfig().SecondsPerSlot) * time.Second
-	// adjusts the genesis time to allow for the clock disparity to
-	// allow for the attestation to be valid.
-	clockAllowance := params.BeaconNetworkConfig().MaximumGossipClockDisparity * 8 / 10
-	newTime := genTime.Add(timePerSlot - clockAllowance)
-	err = validateAggregateAttTime(invalidAttSlot, newTime)
-	if err != nil {
-		t.Errorf("Expected attestation time to be valid, but it was not: %v", err)
-	}
-	// re-determine the current slot
-	currSlot = helpers.SlotsSince(genTime)
-	err = validateAggregateAttTime(currSlot+1, genTime)
-	if err == nil {
-		t.Error("Expected attestation time to be invalid, but it was marked as valid")
-	}
-	err = validateAggregateAttTime(currSlot-10, genTime)
-	if err != nil {
-		t.Errorf("Expected attestation time to be valid, but it was not: %v", err)
 	}
 }
